Refactor: split exact_edt.rs
